0097 chunked チャンクサイズと body_consumed を u64 に統一する (#90)
- BodyChunkedData { remaining } / body_consumed / max_body_size / BodyTooLarge を u64 に変更する
- chunk_size パースを u64::from_str_radix に変更する
- u32::MAX 超チャンクサイズの単体テストを追加する

+++ fuzz/fuzz_targets/fuzz_decoder_limits.rs
@@ -25,7 +25,7 @@ fn build_limits(input: &FuzzLimits) -> DecoderLimits {
         max_buffer_size: input.max_buffer_size as usize,
         max_headers_count: input.max_headers_count as usize,
         max_header_line_size: input.max_header_line_size as usize,
-        max_body_size: input.max_body_size as usize,
+        max_body_size: input.max_body_size as u64,
         max_chunk_line_size: 64,
     }
 }
